Moved display.rb to the bin folder
