+++ bin/display.rb
